Commit a proof of concept for automating bluetoothctl commands for resetting Bluetooth connections.
This commit is a proof of concept and needs to be cleaned up. bluetoothctl commands are called as separate subprocesses, but the combination of commands should be called from within a class.
TODO Move functions to a class and keep a single bluetoothctl subprocess open while streaming commands rather than closing subprocesses after each Linux command.